Simplified and moved tests to appopriate space
